==> Notebooks/1_flights_data_collection.ipynb <==
# Data Gathering
Gather the information for the flights and flights_test tables

import psycopg2

import pandas as pd

con = psycopg2.connect()
print("Database opened successfully")

cur.execute("SELECT fl_date FROM flights")

***Collect the flights table***
- Extract by carrier and export to csv

df_dl = pd.read_sql("""SELECT *
                        FROM flights
                        WHERE mkt_unique_carrier = 'DL' """, con
                    )

df_dl.to_csv("carrier_dl.csv", index = False)

df.mkt_unique_carrier.unique()

df_WN = pd.read_sql("""SELECT *
                        FROM flights
                        WHERE mkt_unique_carrier = 'WN' """, con
                    )

df_WN.to_csv("carrier_WN.csv", index=False)

df_UA = pd.read_sql("""SELECT *
                        FROM flights
                        WHERE mkt_unique_carrier = 'UA' """, con
                    )

df_UA.to_csv("carrier_UA.csv", index=False)

df_G4 = pd.read_sql("""SELECT *
                        FROM flights
                        WHERE mkt_unique_carrier = 'G4' """, con
                    )

df_G4.to_csv("carrier_G4.csv", index = False)

df_AA = pd.read_sql("""SELECT *
                        FROM flights
                        WHERE mkt_unique_carrier = 'AA' """, con
                    )

df_AA.to_csv("carrier_AA.csv", index = False)

df_NK = pd.read_sql("""SELECT *
                        FROM flights
                        WHERE mkt_unique_carrier = 'NK' """, con
                    )

df_NK.to_csv("carrier_NK.csv", index = False)

df_HA = pd.read_sql("""SELECT *
                        FROM flights
                        WHERE mkt_unique_carrier = 'HA' """, con
                    )

df_HA.to_csv("carrier_HA.csv" , index = False)

df_F9 = pd.read_sql("""SELECT *
                        FROM flights
                        WHERE mkt_unique_carrier = 'F9' """, con
                    )

df_F9.to_csv("carrier_F9.csv", index = False)

df_AS = pd.read_sql("""SELECT *
                        FROM flights
                        WHERE mkt_unique_carrier = 'AS' """, con
                    )

df_AS.to_csv("carrier_AS.csv", index = False)

df_B6 = pd.read_sql("""SELECT *
                        FROM flights
                        WHERE mkt_unique_carrier = 'B6' """, con
                    )

df_B6.to_csv("carrier_B6.csv", index = False)

df_VX = pd.read_sql("""SELECT *
                        FROM flights
                        WHERE mkt_unique_carrier = 'VX' """, con
                    )

df_VX.to_csv("carrier_VX.csv", index=False)

df = pd.read_sql("""SELECT mkt_unique_carrier
                        FROM flights""", con
                    )

df = pd.read_sql("""SELECT mkt_unique_carrier, COUNT(fl_date)
                        FROM flights_test 
                        GROUP BY mkt_unique_carrier""", con
                    )

df

***Collect the flights table***
- Extract by carrier and export to csv

df_test_AA = pd.read_sql("""SELECT *
                            FROM flights_test
                            WHERE mkt_unique_carrier = 'AA'""",con)

df_test_AA.to_csv("AA_test.csv", index=False)

df_test_AS = pd.read_sql("""SELECT *
                            FROM flights_test
                            WHERE mkt_unique_carrier = 'AS'""",con)

df_test_AS.to_csv("AS_test.csv", index=False)

df_test_B6 = pd.read_sql("""SELECT *
                            FROM flights_test
                            WHERE mkt_unique_carrier = 'B6'""",con)

df_test_B6.to_csv("B6_test.csv", index=False)

df_test_DL = pd.read_sql("""SELECT *
                            FROM flights_test
                            WHERE mkt_unique_carrier = 'DL'""",con)

df_test_DL.to_csv("DL_test.csv", index=False)

df_test_F9 = pd.read_sql("""SELECT *
                            FROM flights_test
                            WHERE mkt_unique_carrier = 'F9'""",con)

df_test_F9.to_csv("F9_test.csv", index=False)

df_test_G4 = pd.read_sql("""SELECT *
                            FROM flights_test
                            WHERE mkt_unique_carrier = 'G4'""",con)

df_test_G4.to_csv("G4_test.csv", index=False)

df_test_HA = pd.read_sql("""SELECT *
                            FROM flights_test
                            WHERE mkt_unique_carrier = 'HA'""",con)

df_test_HA.to_csv("HA_test.csv", index=False)

df_test_NK = pd.read_sql("""SELECT *
                            FROM flights_test
                            WHERE mkt_unique_carrier = 'NK'""",con)

df_test_NK.to_csv("NK_test.csv", index=False)

df_test_UA = pd.read_sql("""SELECT *
                            FROM flights_test
                            WHERE mkt_unique_carrier = 'UA'""",con)

df_test_UA.to_csv("UA_test.csv", index=False)

df_test_WN = pd.read_sql("""SELECT *
                            FROM flights_test
                            WHERE mkt_unique_carrier = 'WN'""",con)

df_test_WN.to_csv("WN_test.csv", index=False)

df_test_VX = pd.read_sql("""SELECT *
                            FROM flights_test
                            WHERE mkt_unique_carrier = 'VX'""",con)

***Complete Preliminary Data Exploration***
- work with a single carrier
- explore null values

df_WN.shape

df_WN.columns

temp = df_VX[df_VX.weather_delay.isnull()==True]

tmep = temp[["dep_delay", "arr_delay", "cancelled","carrier_delay", "weather_delay", "nas_delay", "security_delay", "late_aircraft_delay"]]

t_can = tmep[tmep.cancelled == 1]
t_not = tmep[tmep.cancelled == 0]

t_not_out = t_not.sort_values(by = "dep_delay", ascending = False)

t_not_out.head(15)

print("Max Not Cancelled dep_delay: ", t_not.dep_delay.max())
print("Max Not Cancelled arr_delay: ", t_not.arr_delay.max())
print("Max Cancelled dep_delay: ", t_can.dep_delay.max())
print("Max Cancelled arr_delay: ", t_can.arr_delay.max())

temp = df_B6[df_B6.weather_delay.isnull()==False]

tmep = temp[["dep_delay", "arr_delay", "cancelled","carrier_delay", "weather_delay", "nas_delay", "security_delay", "late_aircraft_delay"]]

t_not = tmep[tmep.cancelled == 0]

t_not.sort_values(by="dep_delay")

tt.cancelled.unique()

def delete_no_name (df):
    df = df.drop("no_name", axis=1)
    return df

df_AA = delete_no_name(df_AA)
df_AA.head()

df_AS = delete_no_name(df_AS)
df_B6 = delete_no_name(df_B6)
df_DL = delete_no_name(df_dl)
df_F9 = delete_no_name(df_F9)
df_G4 = delete_no_name(df_G4)
df_HA = delete_no_name(df_HA)
df_NK = delete_no_name(df_NK)
df_UA = delete_no_name(df_UA)
df_WN = delete_no_name(df_WN)

df_DL.to_csv("carrier_dl.csv", index = False)
df_AA.to_csv("carrier_AA.csv", index = False)
df_AS.to_csv("carrier_AS.csv", index = False)
df_B6.to_csv("carrier_B6.csv", index = False)
df_F9.to_csv("carrier_F9.csv", index = False)
df_G4.to_csv("carrier_G4.csv", index = False)
df_HA.to_csv("carrier_HA.csv", index = False)
df_NK.to_csv("carrier_NK.csv", index = False)
df_UA.to_csv("carrier_UA.csv", index = False)
df_WN.to_csv("carrier_WN.csv", index = False)

df_VX = delete_no_name(df_VX)
df_VX.to_csv("carrier_VX.csv", index = False)

df_test_WN.info()

df_WN[df_WN.arr_delay > 14]
