==> Notebooks/1_flights_null_treatment.ipynb <==
## Flights Null Treatment

Notebook for the exploration of the null values and the null value treatment

---
**INPUT** raw flights data
**OUTPUT** flights nulls removed

---

import pandas as pd
import math

WN = pd.read_csv("carrier_WN.csv")

WN.columns

WN.tail_num.fillna("EMPTY", inplace=True)

WN.dep_delay.fillna(0, inplace=True)

WN.taxi_out.fillna(0, inplace=True)

WN.wheels_off.fillna(0, inplace=True)

def check_collection_drop(col1, col2):
    if math.isnan(col1):
        if col2 == 1:
            return 0
    return col1 

WN.wheels_on = WN.apply(lambda x: check_collection_drop(x.wheels_on, x.cancelled),axis=1)
WN = WN[WN.wheels_on.notnull()]

WN.taxi_in = WN.apply(lambda x: check_collection_drop(x.taxi_in, x.cancelled),axis=1)
WN = WN[WN.taxi_in.notnull()]

WN.arr_time = WN.apply(lambda x: check_collection_drop(x.arr_time, x.cancelled),axis=1)
WN = WN[WN.arr_time.notnull()]

WN.arr_delay = WN.apply(lambda x: check_collection_drop(x.arr_delay, x.cancelled),axis=1)
WN = WN[WN.arr_delay.notnull()]

WN.cancellation_code.fillna(0, inplace=True)

WN.actual_elapsed_time = WN.apply(lambda x: check_collection_drop(x.actual_elapsed_time, x.cancelled),axis=1)
WN = WN[WN.actual_elapsed_time.notnull()]

WN.air_time = WN.apply(lambda x: check_collection_drop(x.air_time, x.cancelled),axis=1)
WN = WN[WN.air_time.notnull()]

def check_arr_delay (col1, col2):
    if math.isnan(col1):
        if col2 < 15:
            return 0
    return col1

WN.shape

WN.carrier_delay= WN.apply(lambda x: check_arr_delay(x.carrier_delay, x.arr_delay),axis=1)
WN = WN[WN.carrier_delay.notnull()]

WN.shape

WN.weather_delay= WN.apply(lambda x: check_arr_delay(x.weather_delay, x.arr_delay),axis=1)
WN.nas_delay= WN.apply(lambda x: check_arr_delay(x.nas_delay, x.arr_delay),axis=1)
WN.security_delay= WN.apply(lambda x: check_arr_delay(x.security_delay, x.arr_delay),axis=1)
WN.late_aircraft_delay= WN.apply(lambda x: check_arr_delay(x.late_aircraft_delay, x.arr_delay),axis=1)

def replace_with_column (col1, col2):
    if math.isnan(col1):
        return col2
    return col1

WN.dep_time= WN.apply(lambda x: replace_with_column(x.dep_time, x.crs_dep_time),axis=1)

WN.first_dep_time= WN.apply(lambda x: replace_with_column(x.first_dep_time, x.dep_time),axis=1)

WN.total_add_gtime.fillna(0, inplace=True)
WN.longest_add_gtime.fillna(0, inplace=True)

WN.to_csv("carrier_WN_nullsremoved.csv", index=False)

## Repeat for Remaining Carriers

def remove_nulls (WN):
    WN.tail_num.fillna("EMPTY", inplace=True)
    WN.dep_delay.fillna(0, inplace=True)
    WN.taxi_out.fillna(0, inplace=True)
    WN.wheels_off.fillna(0, inplace=True)
    WN.cancellation_code.fillna(0, inplace=True)
    WN.wheels_on = WN.apply(lambda x: check_collection_drop(x.wheels_on, x.cancelled),axis=1)
    WN = WN[WN.wheels_on.notnull()]
    WN.taxi_in = WN.apply(lambda x: check_collection_drop(x.taxi_in, x.cancelled),axis=1)
    WN = WN[WN.taxi_in.notnull()]
    WN.arr_time = WN.apply(lambda x: check_collection_drop(x.arr_time, x.cancelled),axis=1)
    WN = WN[WN.arr_time.notnull()]
    WN.arr_delay = WN.apply(lambda x: check_collection_drop(x.arr_delay, x.cancelled),axis=1)
    WN = WN[WN.arr_delay.notnull()]
    WN.actual_elapsed_time = WN.apply(lambda x: check_collection_drop(x.actual_elapsed_time, x.cancelled),axis=1)
    WN = WN[WN.actual_elapsed_time.notnull()]
    WN.air_time = WN.apply(lambda x: check_collection_drop(x.air_time, x.cancelled),axis=1)
    WN = WN[WN.air_time.notnull()]
    WN.carrier_delay= WN.apply(lambda x: check_arr_delay(x.carrier_delay, x.arr_delay),axis=1)
    WN = WN[WN.carrier_delay.notnull()]
    WN.weather_delay= WN.apply(lambda x: check_arr_delay(x.weather_delay, x.arr_delay),axis=1)
    WN.nas_delay= WN.apply(lambda x: check_arr_delay(x.nas_delay, x.arr_delay),axis=1)
    WN.security_delay= WN.apply(lambda x: check_arr_delay(x.security_delay, x.arr_delay),axis=1)
    WN.late_aircraft_delay= WN.apply(lambda x: check_arr_delay(x.late_aircraft_delay, x.arr_delay),axis=1)
    WN.dep_time= WN.apply(lambda x: replace_with_column(x.dep_time, x.crs_dep_time),axis=1)
    WN.first_dep_time= WN.apply(lambda x: replace_with_column(x.first_dep_time, x.dep_time),axis=1)
    WN.total_add_gtime.fillna(0, inplace=True)
    WN.longest_add_gtime.fillna(0, inplace=True)
    return WN

AA = pd.read_csv("carrier_AA.csv")
print(AA.shape)
AA = remove_nulls(AA)
print(AA.shape)
WN.to_csv("carrier_AA_nullsremoved.csv", index=False)

AA.to_csv("carrier_AA_nullsremoved.csv", index=False)

AS = pd.read_csv("carrier_AS.csv")
AS = remove_nulls(AS)
AS.to_csv("carrier_AS_nullsremoved.csv", index=False)

B6 = pd.read_csv("carrier_B6.csv")
B6 = remove_nulls(B6)
B6.to_csv("carrier_B6_nullsremoved.csv", index=False)

DL = pd.read_csv("carrier_dl.csv")
DL = remove_nulls(DL)
DL.to_csv("carrier_DL_nullsremoved.csv", index=False)

F9 = pd.read_csv("carrier_F9.csv")
F9 = remove_nulls(F9)
F9.to_csv("carrier_F9_nullsremoved.csv", index=False)

G4 = pd.read_csv("carrier_G4.csv")
G4 = remove_nulls(G4)
G4.to_csv("carrier_G4_nullsremoved.csv", index=False)

HA = pd.read_csv("carrier_HA.csv")
HA = remove_nulls(HA)
HA.to_csv("carrier_HA_nullsremoved.csv", index=False)

NK = pd.read_csv("carrier_NK.csv")
NK = remove_nulls(NK)
NK.to_csv("carrier_NK_nullsremoved.csv", index=False)

UA = pd.read_csv("carrier_UA.csv")
UA = remove_nulls(UA)
UA.to_csv("carrier_UA_nullsremoved.csv", index=False)

VX = pd.read_csv("carrier_VX.csv")
VX = remove_nulls(VX)
VX.to_csv("carrier_VX_nullsremoved.csv", index=False)
