==> Notebooks/1_passengers_fuel_preprocess.ipynb <==
# Connect to the Databases

import psycopg2
import pandas as pd

# Connection parameters
param_dic = {
    "host"      : "mid-term-project.ca2jkepgjpne.us-east-2.rds.amazonaws.com",
    "database"  : "mid_term_project",
    "port"      : "5432",
    "user"      : "lhl_student",
    "password"  : "lhl_student"
}

# function to create the connexion 
def connect(params_dic):
    """ Connect to the PostgreSQL database server """
    conn = None
    try:
        # connect to the PostgreSQL server
        print('Connecting to the PostgreSQL database...')
        conn = psycopg2.connect(**params_dic)
    except (Exception, psycopg2.DatabaseError) as error:
        print(error)
        sys.exit(1) 
    print("Connection successful")
    return conn

# function to convert SQL DB to a panda DF
def postgsql_to_df(conn, select_query, column_names):
    """
    Tranform a SELECT query into a pandas dataframe
    """
    cursor = conn.cursor()
    try:
        cursor.execute(select_query)
    except (Exception, psycopg2.DatabaseError) as error:
        print("Error: %s" % error)
        cursor.close()
        return 1
    
    # Naturally we get a list of tupples
    tupples = cursor.fetchall()
    cursor.close()
    # We just need to turn it into a pandas dataframe
    df = pd.DataFrame(tupples, columns=column_names)
    return df

### Create a panda fuel db

col_names_fuel = ['month', 'airline_id', 'unique_carrier', 'carrier', 'carrier_name', 'carrier_group_new', 'sdomt_gallons', 'satl_gallons', 'spac_gallons', 'slat_gallons', 'sint_gallons', 'ts_gallons', 'tdomt_gallons', 'tint_gallons', 'total_gallons','sdomt_cost', 'satl_cost', 'spac_cost', 'slat_cost', 'sint_cost', 'ts_cost', 'tdomt_cost', 'tint_cost', 'total_cost', 'year']
fuel = postgsql_to_df(connect(param_dic), "select * from fuel_comsumption", col_names_fuel)

fuel.head()

fuel.shape

### Create a panda passengers db
Issue with column numbers. Have 37 in documentaiton but 38 in the data. 
- See if there a column with Nan : no
- See if there is a dupplicated column: yes, 

col_names_passengers = ['unknown', 'departures_scheduled', 'departures_performed', 'payload', 'seats', 'passengers', 'freight', 'mail', 'distance', 'ramp_to_ramp', 'air_time', 'unique_carrier', 'airline_id', 'unique_carrier_name', 'region', 'carrier','carrier_name', 'carrier_group', 'carrier_group_new', 'origin_airport_id', 'origin_city_market_id', 'origin', 'origin_city_name', 'origin_country', 'origin_country_name', 'dest_airport_id', 'dest_city_market_id', 'dest', 'dest_city_name', 'dest_country', 'dest_country_name', 'aircraft_group', 'aircraft_type', 'aircraft_config', 'month', 'year', 'distance_group', 'class']
passengers = postgsql_to_df(connect(param_dic), "select * from passengers", col_names_passengers)

passengers.head()

passengers.shape

print(len(col_names_passengers)) # added an 'unknown' column
print(len(passengers.columns))

#explore 

passengers.iloc[:5,:10]

passengers.iloc[:5,10:20]

passengers.iloc[:5,20:30]

passengers.iloc[:5,30:]

# check if we have a column with only zero values. No
db_length = passengers.shape[0]
for column in list(passengers.columns):
    if passengers[column][passengers[column] == 0].count() == db_length:
        print(column)

# see if there are any dupplicate columns. We have 2350497 lines in the database

# comapre unknown and departures_scheduled
print(passengers.unknown[passengers.unknown == passengers.departures_scheduled].count())

# comapre payload and seats
print(passengers.unknown[passengers.payload == passengers.seats].count())

# comapre payload and freight
print(passengers.unknown[passengers.payload == passengers.freight].count())

# comapre seats and freight
print(passengers.unknown[passengers.seats == passengers.freight].count())

# comapre airline_id and carrier Those look extremely similar!
print(passengers.unknown[passengers.airline_id == passengers.carrier].count())

# comapre carrier_name and carrier_group
print(passengers.unknown[passengers.carrier_name == passengers.carrier_group].count())

# calculate % similarity between airline_id and carrier
passengers.unknown[passengers.airline_id == passengers.carrier].count()/db_length*100

passengers.iloc[:5,[12,15]]

# actually they both belong! They are carrier_name and unique_carrier_name

# rename dataframe with right column names
col_names_passengers = ['departures_scheduled', 'departures_performed', 'payload', 'seats', 'passengers', 'freight', 'mail', 'distance', 'ramp_to_ramp', 'air_time', 'unique_carrier', 'airline_id', 'unique_carrier_name', 'region', 'carrier','carrier_name', 'carrier_group', 'carrier_group_new', 'origin_airport_id', 'origin_city_market_id', 'origin', 'origin_city_name', 'origin_country', 'origin_country_name', 'dest_airport_id', 'dest_city_market_id', 'dest', 'dest_city_name', 'dest_country', 'dest_country_name', 'aircraft_group', 'aircraft_type', 'aircraft_config', 'year',  'month', 'day', 'distance_group', 'class']
passengers.columns = col_names_passengers

passengers.head()

# Explore the databases

## Passengers

passengers.info()

### Check for dupplicates

passengers.duplicated().sum()

### Check for nulls

# check the percentage of null values per variable
passengers.isnull().sum()

# drop rows with missing values
print(len(passengers))
passengers = passengers.dropna()
print(len(passengers))

## Fuel

# check for duplicates
fuel.duplicated().sum()

# check the percentage of null values per variable
(fuel.isnull().sum()/fuel.shape[0])*100

# drop rows with missing values
print(len(fuel))
fuel = fuel.dropna()
print(len(fuel))

# Export Databases into CSV

# passengers
passengers.to_csv("passengers.csv")

# fuel consumption
fuel.to_csv("fuel.csv")
