==> Notebooks/2_ExtractSample.ipynb <==
### Extract Sample from Flights

import pandas as pd

result = pd.read_csv("flights_nullsremoved.csv")

result.shape

sample1000 = result.sample(1000)

sample1000.shape

sample5000000 = result.sample(5000000)

sample1000.to_csv("flight_sample_small.csv", index=False)
sample5000000.to_csv("flight_sample_large.csv", index=False)
