==> Notebooks/2_fuel_EDA.ipynb <==
import seaborn as sns
import matplotlib.pyplot as plt
%matplotlib inline
import pandas as pd
import numpy as np
import seaborn as sns

fuel = pd.read_csv('fuel_cleaned.csv', index_col = 0)

# get shape and length
print(fuel.shape)
db_length = fuel.shape[0]

fuel.info()

# explore all columns
fuel.iloc[:5, :10]

fuel.iloc[:5, 10:20]

fuel.iloc[:5, 20:]

# Numerical data

numeric_features = fuel.select_dtypes(include=[np.number])
numeric_features.columns

# see if variables are categorical / continuous: look at number of unique values for each feature
for column in numeric_features.columns:
    print(column, len(fuel[column].unique()))

# String data

string_data = fuel.dtypes[fuel.dtypes == 'object'].index
string_data

# Drop redundant features

# redundant: keeping unique_carrier

fuel = fuel.drop('carrier', axis = 1)

# Other EDA

# Check if year/month/airlineID is the unique key of this db

len(fuel.groupby(['year', 'month', 'airline_id']).count())

len(fuel.groupby(['airline_id']).count())
