==> Notebooks/2_passengers_EDA.ipynb <==
import seaborn as sns
import matplotlib.pyplot as plt
%matplotlib inline
import pandas as pd
import numpy as np
import seaborn as sns

passengers = pd.read_csv('passengers_cleaned.csv', index_col = 0)

# get shape and length
print(passengers.shape)
db_length = passengers.shape[0]

passengers.info()

# explore all columns
passengers.iloc[:5,:10]

passengers.iloc[:5,10:20]

passengers.iloc[:5,20:30]

passengers.iloc[:5,30:]

# Redundant features

# unique carrier and carrier
passengers.carrier[passengers.carrier == passengers.unique_carrier].count()/db_length*100

# unique_carrier_name and carrier_name
passengers.carrier_name[passengers.unique_carrier_name == passengers.carrier_name].count()/db_length*100

# carrier_group and carrier_group_new
passengers.carrier_group[passengers.carrier_group == passengers.carrier_group_new].count()/db_length*100

# Highly correlated features

# origin and origin_city_name

# create placeholder database for two features
df_correlation_check = pd.DataFrame(columns = [1,2])

# first feature
df_correlation_check[1] = passengers.origin.astype('category').cat.codes

# second feature
df_correlation_check[2] = passengers.origin_city_name.astype('category').cat.codes

# check correlation
df_correlation_check.corr()

# origin_country and origin_country_name

# create placeholder database for two features
df_correlation_check = pd.DataFrame(columns = [1,2])

# first feature
df_correlation_check[1] = passengers.origin_country.astype('category').cat.codes

# second feature
df_correlation_check[2] = passengers.origin_country_name.astype('category').cat.codes

# check correlation
df_correlation_check.corr()

# dest and dest_city_name

# create placeholder database for two features
df_correlation_check = pd.DataFrame(columns = [1,2])

# first feature
df_correlation_check[1] = passengers.dest.astype('category').cat.codes

# second feature
df_correlation_check[2] = passengers.dest_city_name.astype('category').cat.codes

# check correlation
df_correlation_check.corr()

# dest_country and dest_country_name

# create placeholder database for two features
df_correlation_check = pd.DataFrame(columns = [1,2])

# first feature
df_correlation_check[1] = passengers.dest_country.astype('category').cat.codes

# second feature
df_correlation_check[2] = passengers.dest_country_name.astype('category').cat.codes

# check correlation
df_correlation_check.corr()

# unique_carrier and unique_carrier_name

# create placeholder database for two features
df_correlation_check = pd.DataFrame(columns = [1,2])

# first feature
df_correlation_check[1] = passengers.unique_carrier.astype('category').cat.codes

# second feature
df_correlation_check[2] = passengers.unique_carrier_name.astype('category').cat.codes

# check correlation
df_correlation_check.corr()

# Numerical data

numeric_features = passengers.select_dtypes(include=[np.number])
numeric_features.columns

# see if variables are categorical / continuous: look at number of unique values for each feature
for column in numeric_features.columns:
    print(column, len(passengers[column].unique()))

# print categories for categorical data
print('carrier_group', passengers.carrier_group.unique())
print('carrier_group_new', passengers.carrier_group.unique())
print('aircraft_config', passengers.carrier_group.unique())

# String data

string_data = passengers.dtypes[passengers.dtypes == 'object'].index
string_data

print('unique_carrier (length)', len(passengers.unique_carrier.unique()))
print('unique_carrier_name (length)', len(passengers.unique_carrier_name.unique()))
print('region', passengers.region.unique())
print('carrier (length)', len(passengers.carrier.unique()))
print('carrier_name (length)', len(passengers.carrier_name.unique()))
print('origin (length)', len(passengers.origin.unique()))
print('origin_country (length)', len(passengers.origin_country.unique()))
print('class', passengers['class'].unique())

# Drop redundant features

# redundant: keeping unique_carrier, unique_carrier_name, carrier_group_new

passengers = passengers.drop(['carrier', 'carrier_name', 'carrier_group'], axis = 1)
