==> Notebooks/3_Training_Prep.ipynb <==
## Training Prep

The purpose of this notebook is to provide scaled data for the models we will be using

---

**INPUT**: flights_cleaned <br>
**OUTPUT**: flights_scaled and scaler.pkl <br>
**Next Steps**: PCA (where required) and/or Modeling

---

### Import Packages

import pandas as pd
import numpy as np
from sklearn.preprocessing import StandardScaler
from pickle import dump

---

### Import Cleaned Data

df = pd.read_csv("flights_cleaned.csv")

df.head(2)

df.drop(columns = ['origin_city_name', 'dest_city_name'], inplace = True)

df.columns

### Work with our data seperate from our targets

y = df[['arr_delay', 'cancelled', 'delay_type']]

X = df.drop(columns = ['arr_delay', 'cancelled', 'delay_type']).values

scaler = StandardScaler()

scaler.fit(X)

X_scaled = scaler.transform(X)

*Export scaler to use on evaluation data*

#save the scaler
dump(scaler, open('scaler.pkl', 'wb'))

#prep train and test files
trainX = pd.DataFrame(data = X_scaled,
              columns = ['branded_code_share', 'origin', 'dest', 'crs_dep_time', 'crs_arr_time',
                         'crs_elapsed_time', 'air_time', 'distance','fl_month', 'fl_day_of_week', 
                         'fl_week_of_month', 'mkt_op_combo','fl_type', 'm_hist_dep_delay',
                         'med_hist_dep_delay', 'm_hist_arr_delay','med_hist_arr_delay'])

train = trainX.merge(y, right_index=True, left_index=True)

train.shape

train.to_csv("flights_scaled.csv", index=False)

train_samp = train.sample(10000)

train_samp.to_csv("flights_scaled_sample.csv", index=False)
