==> Notebooks/4_sample_creation.ipynb <==
# The goal of this file is to create smaller samples for each ML task in order to spot check different algorythms
## INPUTS 
- flights_scaled_sample.csv
- flights_pca_sample.csv  
## OUTPUTS
- db_binary_sample.csv
- db_multiclass_sample
- db_regression_sample
- db_binary_pca_sample.csv
- db_multiclass_pca_sample
- db_regression_pca_sample

import pandas as pd

### From Flights_scaled

flights= pd.read_csv('flights_scaled_sample.csv')

flights.shape

flights.info()

# see portion of delayed flights that have a cause (33% all flights are delayed, 17% have a cause reported)
print(len(flights[flights.arr_delay >0]))
print(len(flights[flights.delay_type >0]))

# get database for binary classification. drop target variables used in other models
db_binary_sample = flights.drop(['arr_delay', 'delay_type'], axis=1)

# get database for multiclass classification - only flights that were delayed, drop target variables used in other models, take only >15min dalays 
db_multiclass_sample = flights[flights['arr_delay'] > 15].drop(['arr_delay', 'cancelled'], axis=1)

# get database for regression. drop target variables used in other models
db_regression_sample = flights.drop(['delay_type', 'cancelled'], axis=1)

# print final shapes
print(db_binary_sample.shape)
print(db_multiclass_sample.shape)
print(db_regression_sample.shape)

# export samples

db_binary_sample.to_csv("db_binary_sample.csv")
db_multiclass_sample.to_csv("db_multiclass_sample.csv")
db_regression_sample.to_csv("db_regression_sample.csv")

### From Flights_pca 

flights_pca = pd.read_csv('flights_pca_sample.csv')

flights_pca.shape

flights_pca.info()

flights_pca.delay_type.unique()

# get database for binary classification. drop target variables used in other models, take only >15min delays
db_binary_pca_sample = flights_pca.drop(['arr_delay', 'delay_type'], axis=1)
len(db_binary_pca)

# get database for multiclass classification - only flights that were delayed, drop target variables used in other models
db_multiclass_pca_sample = flights_pca[flights_pca['arr_delay'] > 15].drop(['arr_delay', 'cancelled'], axis=1)
len(db_multiclass_pca)

# get database for regression. drop target variables used in other models 
db_regression_pca_sample = flights_pca.drop(['delay_type', 'cancelled'], axis=1)
len(db_regression_pca)

# print final shapes

print(db_binary_pca_sample.shape)
print(db_multiclass_pca_sample.shape)
print(db_regression_pca_sample.shape)

# export samples

db_binary_pca_sample.to_csv("db_binary_pca_sample.csv")
db_multiclass_pca_sample.to_csv("db_multiclass_pca_sample.csv")
db_regression_pca_sample.to_csv("db_regression_pca_sample.csv")
